Room service

+++ src/main/java/com/sda/transporeon/finalconferenceroom/reservation/model/ReservationResponse.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class ReservationResponse {
 
-    private Integer reservationId;
+    private Long reservationId;
     private String reservationIdentifier;
     private LocalDateTime reservationStartDate;
     private LocalDateTime ReservationEndDate;
